requirements_final.txt
aiohttp==3.9.5
aiosignal==1.4.0
async-timeout==4.0.3
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.3.2
cryptography==42.0.8
cycler==0.12.1
filelock==3.20.0
filterpy==1.4.5
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2025.12.0
idna==3.11
Jinja2==3.1.6
keyboard==0.13.5
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
MouseInfo==0.1.3
mpmath==1.3.0
mss==10.1.0
multidict==6.7.0
networkx==3.4.2
numpy==1.24.4
opencv-python==4.9.0.80
packaging==25.0
pandas==2.3.3
pillow==10.2.0
propcache==0.4.1
psutil==5.9.8
py-cpuinfo==9.0.0
PyAutoGUI==0.9.54
pycparser==2.23
pycryptodome==3.21.0
pygame==2.5.2
PyGetWindow==0.0.9
PyJWT==2.8.0
PyMsgBox==2.0.1
pyparsing==3.2.5
pyperclip==1.11.0
pypresence==4.3.0
PyRect==0.2.0
PyScreeze==1.0.1
pyserial==3.5
PySide6==6.5.2
PySide6-Addons==6.5.2
PySide6-Essentials==6.5.2
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2025.2
pywin32==306
PyYAML==6.0.3
requests==2.31.0
scipy==1.15.3
seaborn==0.13.2
shiboken6==6.5.2
six==1.17.0
sympy==1.14.0
termcolor==2.4.0
torch==2.4.1
torch-directml==0.2.5.dev240914
torchvision==0.19.1
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.2.65
ultralytics-thop==2.0.18
urllib3==2.2.3
vgamepad==0.1.0
WMI==1.5.1
yarl==1.22.0
